store-api/store-api/requirements.txt
Django~=3.2
environs~=8.0.0
Pillow==9.4.0
asgiref
psycopg2-binary==2.9.5
gunicorn==20.1.0
djangorestframework==3.14.0
drf-yasg==1.21.5
django-cors-headers==3.14.0

